Added support for MCP23017 based encoders & switches
Uses WiringPi ISRs on two GPIO pins that report on pin changes on
the A & B banks of the MCP23017.  When the MCP23017 detects a pin
change on a given bank, it sets the associated INT pin high.  A
WiringPi detects the pin change to high and calls the bank's ISR.
The ISR reads the state of all the pins in the bank and compares
the bits of that bank to the associated bits of the switches and
encoders.  If any of the read bits differs from the last read set
then the update_zynswitch or update_zyncoder function is called.
C #ifdefs are used to avoid additional unnecessary digitalReads
on the I2C bus and the previously read bank registers are passed
into the update_zynswitch & update_zyncoder functions.

Currently, the MCP23017 INTA is connected to WiringPi pin 27 and
INTB to WiringPi pin 25.  Note, that pin 25 is also the infrared
pin on SainSmart Pifi DAC+ board, but this does not interfere.

To setup the correct pins in the webconf for zynthian, under
wiring, set the Wiring Layout to CUSTOM and the pins as:
Encoder A GPIO-pins: 100,103,108,111
Encoder B GPIO-pins: 101,104,109,112
Switch GPIO-pins: 102,105,110,113

The zyncoder_test has been updated for this configuration.